Handle validation and trigger AFTER_VALIDATION_EVENT in the right order
The order of those was different for a promise validation vs a plain
function validation. The trigger was called *before* the validation
handler when the validation is a promise.
Since the errors are added to the changeset in `_handleValidation` the
trigger for promise validations was called before that error was set.