Don't take cached value in _valueFor
It seems that the [ERRORS] object is used here as a "cache" for the
value. In fact it doesn't give us much benefit as it is not called
multiple times during a single validation.

Using the cache for derived property doesn't work because it is not
properly discarded when any relevant properties are changed.

It seems reasonable to remove the usage of cache to avoid potential
issues and confusions.