Validate changeset getter (#54)
* Add failing test for validating changeset getter

If we need to validate based on derived property, not the raw changes
the only place where we can put a getter to calculate the value of that
property is the changeset itself as it buffers all changes.

This test case illustrates the failure of that scenario.

* Try to get a property from the changeset

If there is a validation based on derived property we should get its
value in _valueFor().

* Don't take cached value in _valueFor

It seems that the [ERRORS] object is used here as a "cache" for the
value. In fact it doesn't give us much benefit as it is not called
multiple times during a single validation.

Using the cache for derived property doesn't work because it is not
properly discarded when any relevant properties are changed.

It seems reasonable to remove the usage of cache to avoid potential
issues and confusions.