Allow a changeset to be created in a validated state (#149)
Adds a new `initValidate` changeset option that can be specified when creating an instance of a changeset. When this option is set to `true`, the `_errors` property of the changeset will be populated with all errors for the current `validatorMap` and `_content`. This will result in the changeset being created in the `isValid = false` state when there are errors upon initialization.